+++ CHANGES.md
@@ -4,6 +4,7 @@
 Changes in 7.8
 --------------
 - Handle `multiple` files attribute in warehouse.js & Inputs.php
+- Add FileUploadMultiple(). Handle `multiple` files attribute for FileUpload() in FileUpload.fnc.php
 
 Changes in 7.7
 --------------